{#39} url 분리
